+++ src/main/java/com/kelco/kamenridercraft/entities/bosses/NewMoleImaginEntity.java
@@ -7,7 +7,6 @@ import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.ai.attributes.AttributeSupplier;
 import net.minecraft.world.entity.ai.attributes.Attributes;
 import net.minecraft.world.entity.monster.Monster;
-import net.minecraft.world.entity.monster.Zombie;
 import net.minecraft.world.level.Level;
 
 public class NewMoleImaginEntity extends BaseHenchmenEntity {
@@ -24,7 +23,7 @@ public class NewMoleImaginEntity extends BaseHenchmenEntity {
 
         return Monster.createMonsterAttributes()
         		.add(Attributes.FOLLOW_RANGE, 135.0D)
-        		.add(Attributes.MOVEMENT_SPEED,(double)0.2F)
+        		.add(Attributes.MOVEMENT_SPEED, 0.2F)
         		.add(Attributes.ATTACK_DAMAGE, 8.0D)
         		.add(Attributes.ARMOR, 2.0D)
         		.add(Attributes.MAX_HEALTH, 60.0D);
