+++ src/main/java/com/kelco/kamenridercraft/entities/bosses/NyelvEntity.java
@@ -2,9 +2,7 @@ package com.kelco.kamenridercraft.entities.bosses;
 
 
 import com.kelco.kamenridercraft.entities.footSoldiers.BaseHenchmenEntity;
-import com.kelco.kamenridercraft.item.Ex_Aid_Rider_Items;
 import com.kelco.kamenridercraft.item.Gavv_Rider_Items;
-import com.kelco.kamenridercraft.item.Zero_One_Rider_Items;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.EquipmentSlot;
 import net.minecraft.world.entity.ai.attributes.AttributeSupplier;
@@ -32,7 +30,7 @@ public class NyelvEntity extends BaseHenchmenEntity {
 
         return Monster.createMonsterAttributes()
                 .add(Attributes.FOLLOW_RANGE, 135.0D)
-                .add(Attributes.MOVEMENT_SPEED,(double)0.3F)
+                .add(Attributes.MOVEMENT_SPEED, 0.3F)
                 .add(Attributes.ATTACK_DAMAGE, 10.0D)
                 .add(Attributes.ARMOR, 3.0D)
                 .add(Attributes.MAX_HEALTH, 200.0D);
