+++ src/main/java/com/kelco/kamenridercraft/entities/bosses/OrgaEntity.java
@@ -1,8 +1,6 @@
 package com.kelco.kamenridercraft.entities.bosses;
 
-import com.kelco.kamenridercraft.ServerConfig;
 import com.kelco.kamenridercraft.block.Rider_Blocks;
-import com.kelco.kamenridercraft.entities.MobsCore;
 import com.kelco.kamenridercraft.entities.footSoldiers.BaseHenchmenEntity;
 import com.kelco.kamenridercraft.item.Faiz_Rider_Items;
 
@@ -64,7 +62,7 @@ public class OrgaEntity extends BaseHenchmenEntity {
 	public static AttributeSupplier.Builder setAttributes() {
 		return Monster.createMonsterAttributes()
         		.add(Attributes.FOLLOW_RANGE, 128.0D)
-        		.add(Attributes.MOVEMENT_SPEED,(double)0.30F)
+        		.add(Attributes.MOVEMENT_SPEED, 0.30F)
         		.add(Attributes.ATTACK_DAMAGE, 5.0D)
         		.add(Attributes.MAX_HEALTH, 100.0D);
      }
