+++ src/main/java/com/kelco/kamenridercraft/entities/bosses/PhantomCrusherEntity.java
@@ -1,13 +1,10 @@
 package com.kelco.kamenridercraft.entities.bosses;
 
 import com.kelco.kamenridercraft.entities.footSoldiers.BaseHenchmenEntity;
-import com.kelco.kamenridercraft.item.Build_Rider_Items;
 import net.minecraft.world.entity.EntityType;
-import net.minecraft.world.entity.EquipmentSlot;
 import net.minecraft.world.entity.ai.attributes.AttributeSupplier;
 import net.minecraft.world.entity.ai.attributes.Attributes;
 import net.minecraft.world.entity.monster.Monster;
-import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.Level;
 
 public class PhantomCrusherEntity extends BaseHenchmenEntity {
@@ -22,7 +19,7 @@ public class PhantomCrusherEntity extends BaseHenchmenEntity {
 
         return Monster.createMonsterAttributes()
                 .add(Attributes.FOLLOW_RANGE, 40.0D)
-                .add(Attributes.MOVEMENT_SPEED,(double)0.45F)
+                .add(Attributes.MOVEMENT_SPEED, 0.45F)
                 .add(Attributes.ATTACK_DAMAGE, 10.0D)
                 .add(Attributes.ARMOR, 3.0D)
                 .add(Attributes.MAX_HEALTH, 90.0D);
