+++ src/main/java/com/kelco/kamenridercraft/entities/bosses/SuperGingaOhEntity.java
@@ -26,7 +26,7 @@ public class SuperGingaOhEntity extends BaseHenchmenEntity {
 
         return Monster.createMonsterAttributes()
         		.add(Attributes.FOLLOW_RANGE, 135.0D)
-        		.add(Attributes.MOVEMENT_SPEED,(double)0.35F)
+        		.add(Attributes.MOVEMENT_SPEED, 0.35F)
         		.add(Attributes.ATTACK_DAMAGE, 10.0D)
         		.add(Attributes.ARMOR, 3.0D)
         		.add(Attributes.MAX_HEALTH, 300.0D);
