+++ src/main/java/com/kelco/kamenridercraft/entities/bosses/ZooousPredatorEntity.java
@@ -19,7 +19,7 @@ public class ZooousPredatorEntity extends BaseHenchmenEntity {
 
         return Monster.createMonsterAttributes()
         		.add(Attributes.FOLLOW_RANGE, 135.0D)
-        		.add(Attributes.MOVEMENT_SPEED,(double)0.4F)
+        		.add(Attributes.MOVEMENT_SPEED, 0.4F)
         		.add(Attributes.ATTACK_DAMAGE, 10.0D)
         		.add(Attributes.ARMOR, 10.0D)
         		.add(Attributes.MAX_HEALTH, 200.0D);
